Fix order cache
The existing cache key does not take into account all parameters, leading to caching errors when e.g. loading page two (you will get a hit on the cache for the first page that was viewed and therefore cached).

This change automatically uses the entire args array to create the key, ensuring uniqueness.